requirements.txt
fastapi
jinja2
python-multipart
python-dotenv
requests
uvicorn